make time series length a command line parameter. include script to generate all training data. update readme

+++ training_data/run_single_batch.sh
@@ -8,23 +8,26 @@
 # # University of Waterloo
 
 
-# Batch number
-batch_num=1
+
+# Get command line parameters
+batch_num=$1 # Batch number (we send each batch to a different CPUs to run in parallel)
+ts_len=$2 # Length of time series to generate for training data
+
+# Other parameters
+bif_max=1 # Number of each type of bifurcation requested (we do 1000 per batch)
 
 
-# Make job-specific directory and move to this directory
-mkdir -p output/batch$batch_num
-cd output/batch$batch_num
+# Make batch-specific directory for storing output
+mkdir -p output/ts_$ts_len/batch$batch_num
+# Move to the batch-specific directory
+cd output/ts_$ts_len/batch$batch_num
 
 # Set up timer
 start=`date +%s`
 
 echo Job released
 
-
 # Define variables that count numbers of bifurcations
-bif_max=2 # Number of each type of bifurcation to generate
-batch_num=1 # ID of batch
 hopf_count=0
 fold_count=0
 branch_count=0
@@ -44,17 +47,22 @@ printf "\n\n"
 
 # Generate a model and output equi.csv, pars.csv
 echo Run gen_model.py
-python3 ../../gen_model.py
+python3 ../../../gen_model.py
 
+# Copy AUTO model and constants files to batch-specific directory
+cp ../../../c.model c.model
+cp ../../../model.f90 model.f90
 
-# Run bifurcation continuation using AUTO, output b.out files for each varied parameter
-# (Make sure runs using Python 2)
-# Copy auto model and constants files to job-specific directory
-cp ../../c.model c.model
-cp ../../model.f90 model.f90
+# Run bifurcation continuation using AUTO and output b.out files for each varied parameter
+# (Make sure AUTO runs using Python 2)
 echo Run run_cont.auto
-auto ../../run_cont.auto
-
+# This should not take more than 10 mins - if it does, cancel the run and create new model
+timeout 600 auto ../../../run_cont.auto
+if [[ $? == 124 ]]
+then
+    echo AUTO cancelled as it took over 10 minutes
+	continue
+fi
 # Remove unnecessary files
 rm -f c.model
 rm -f model.f90
@@ -63,7 +71,7 @@ rm -f model.f90
 # bifurcation point, and output 500 points prior to the transition.
 # Also run a null case where parameters are fixed.
 echo Run stoch_sims.py
-python3 ../../stoch_sims.py $hopf_count $fold_count $branch_count $null_h_count $null_f_count $null_b_count $bif_max $batch_num
+python3 ../../../stoch_sims.py $hopf_count $fold_count $branch_count $null_h_count $null_f_count $null_b_count $bif_max $batch_num $ts_len
 
 
 # Update counting variables
@@ -76,23 +84,24 @@ null_b_count=$(sed '6q;d' output_counts/list_counts.txt)
 null_count=$((null_h_count+null_f_count+null_b_count))
 
 
-# Remove old model, auto and simulation files to save space
+# Remove old model, AUTO and simulation files to save space
 rm -r output_model
 rm -r output_auto
 
 done
 
+
+
 # Convert label data and split into training, test, validation
 echo "Convert data to correct form for training"
-python3 ../../to_traindata.py $bif_max $batch_num
+python3 ../../../to_traindata.py $bif_max $batch_num
 
 # Remove single label files
 rm output_labels/label*
 
-
-# Compute residual dynamics and standard EWS from each simulation
-echo "Compute EWS"
-python3 ../../compute_ews.py $bif_max $batch_num
+# Compute residual dynamics after Lowess smoothing for each time series
+echo "Compute residuals"
+python3 ../../../compute_resids.py $bif_max $batch_num
 
 
 # End timer
@@ -101,6 +110,6 @@ runtime=$((end-start))
 echo "Job successfully finished in time of $runtime" seconds.
 
 # Change back to original directory
-cd ../../
+cd ../../../
